[Marker Styles] Adjust styles for marker resize
