+++ app/src/javascripts/components/InfoWindow.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { infoWindowStyle } from '../styles/mapMarkerStyles';
+import { infoWindowStyle } from '../styles/infoWindowStyles';
 
 const InfoWindow = ({ description }) => {
   return <div style={infoWindowStyle}>{description}</div>;
