+++ app/src/javascripts/styles/mapMarkerStyles.js
@@ -1,6 +1,11 @@
 const iconUrl = require('./svgs/map_icon_std.svg');
 
 const MARKER_DIAMETER = 25;
+const MARKER_WIDTH = 25;
+const MARKER_HEIGHT = 32.5;
+
+const MARKER_HOVER_WIDTH = MARKER_WIDTH * 2;
+const MARKER_HOVER_HEIGHT = MARKER_HEIGHT * 2;
 
 const mapMarkerStyle = {
   // initially any map object has left top corner at lat lng coordinates
@@ -8,8 +13,8 @@ const mapMarkerStyle = {
   position: 'absolute',
   width: 25,
   height: 32.5,
-  left: -MARKER_DIAMETER / 2,
-  top: -MARKER_DIAMETER / 2,
+  left: -MARKER_DIAMETER / 2, // -12.5
+  top: -MARKER_DIAMETER / 2, // -12.5
 
   //borderRadius: MARKER_DIAMETER,
   // backgroundColor: 'red',
@@ -20,20 +25,10 @@ const mapMarkerStyle = {
 
 const mapMarkerHoverStyle = {
   ...mapMarkerStyle,
-  left: MARKER_DIAMETER / 5 * 0.625,
-  top: MARKER_DIAMETER / 5 * 0.625,
-  width: 40,
-  height: 52, // 1.3 x height for this particular icon
+  left: -MARKER_DIAMETER,
+  top: -MARKER_DIAMETER * 2,
+  width: MARKER_HOVER_WIDTH,
+  height: MARKER_HOVER_HEIGHT, // 2 x height for this particular icon
 };
 
-const infoWindowStyle = {
-  position: 'absolute',
-  right: -MARKER_DIAMETER / 2 + 30,
-  bottom: -MARKER_DIAMETER / 2 + 30,
-  width: 70,
-  height: 50,
-  backgroundColor: 'green',
-  zIndex: 1000,
-};
-
-export { mapMarkerStyle, mapMarkerHoverStyle, infoWindowStyle, MARKER_DIAMETER };
+export { mapMarkerStyle, mapMarkerHoverStyle, MARKER_DIAMETER };
